+++ infinite-rs/examples/load_all_modules.rs
@@ -1,8 +1,10 @@
+use bitflags::bitflags;
 use infinite_rs::tag::types::common_types::{
-    AnyTag, FieldBlock, FieldByteFlags, FieldLongFlags, FieldReference, FieldStringId,
+    AnyTag, FieldBlock, FieldByteFlags, FieldCharEnum, FieldLongEnum, FieldReference, FieldStringId,
 };
 use infinite_rs::ModuleFile;
 use infinite_rs_derive::TagStructure;
+use num_enum::TryFromPrimitive;
 
 fn load_modules(deploy_path: String) -> infinite_rs::Result<Vec<ModuleFile>> {
     let mut modules = Vec::new();
@@ -13,10 +15,10 @@ fn load_modules(deploy_path: String) -> infinite_rs::Result<Vec<ModuleFile>> {
         if entry.file_type().is_file() {
             let file_path = entry.path().to_str().unwrap();
             if file_path.ends_with(".module") {
-                let mut module = ModuleFile::new();
-                match module.read(file_path) {
+                let module = ModuleFile::from_path(file_path);
+                match module {
                     Ok(_) => {
-                        modules.push(module);
+                        modules.push(module?);
                         println!("Read module: {}", file_path);
                     }
                     Err(err) => {
@@ -33,12 +35,32 @@ fn load_modules(deploy_path: String) -> infinite_rs::Result<Vec<ModuleFile>> {
 #[derive(Default, Debug, TagStructure)]
 #[data(size(0x30))]
 struct MaterialShaderFunctionParameter {
-    #[data(offset(0x00))]
-    input_type: FieldLongFlags,
     #[data(offset(0x04))]
     input_name: FieldStringId,
-    #[data(offset(0x0C))]
-    output_modifier: FieldByteFlags,
+}
+
+bitflags! {
+    #[derive(Default, Debug)]
+    pub struct MaterialFlags : u8 {
+        const CONVERTED_FROM_SHADER = 0b01;
+        const DECAL_POST_LIGHTING = 0b10;
+        const RUNTIME_GENERATED = 0b100;
+    }
+}
+
+#[derive(TryFromPrimitive, Debug, Default)]
+#[repr(u32)]
+enum MaterialParameterType {
+    #[default]
+    Bitmap,
+    Real,
+    Int,
+    Bool,
+    Color,
+    ScalarGPUProperty,
+    ColorGPUProperty,
+    String,
+    Preset,
 }
 
 #[derive(Default, Debug, TagStructure)]
@@ -46,6 +68,8 @@ struct MaterialShaderFunctionParameter {
 struct MaterialParameter {
     #[data(offset(0x8))]
     bitmap: FieldReference,
+    #[data(offset(0x4))]
+    parameter_type: FieldLongEnum<MaterialParameterType>,
     #[data(offset(0x80))]
     function_parameters: FieldBlock<MaterialShaderFunctionParameter>,
 }
@@ -64,6 +88,53 @@ struct PostProcessDefinition {
     textures: FieldBlock<MaterialPostprocessTexture>,
 }
 
+#[derive(TryFromPrimitive, Debug, Default)]
+#[repr(u8)]
+enum MaterialStyleShaderSupportedLayers {
+    #[default]
+    Supports1Layer,
+    Supports4Layers,
+    Supports7Layers,
+    LayerShaderDisabled,
+}
+
+#[derive(TryFromPrimitive, Debug, Default)]
+#[repr(u8)]
+enum MaterialStyleShaderSupportsDamageEnum {
+    #[default]
+    No,
+    Yes,
+}
+
+#[derive(Default, Debug, TagStructure)]
+#[data(size(0x5c))]
+struct MaterialStyleInfo {
+    #[data(offset(0x00))]
+    material_style: FieldReference,
+    #[data(offset(0x1C))]
+    material_style_tag: FieldReference,
+    #[data(offset(0x38))]
+    region_name: FieldStringId,
+    #[data(offset(0x3C))]
+    base_intention: FieldStringId,
+    #[data(offset(0x40))]
+    mask0_red_channel_intention: FieldStringId,
+    #[data(offset(0x44))]
+    mask0_green_channel_intention: FieldStringId,
+    #[data(offset(0x48))]
+    mask0_blue_channel_intention: FieldStringId,
+    #[data(offset(0x4C))]
+    mask1_red_channel_intention: FieldStringId,
+    #[data(offset(0x50))]
+    mask1_green_channel_intention: FieldStringId,
+    #[data(offset(0x54))]
+    mask1_blue_channel_intention: FieldStringId,
+    #[data(offset(0x58))]
+    supported_layers: FieldCharEnum<MaterialStyleShaderSupportedLayers>,
+    #[data(offset(0x59))]
+    requires_damage: FieldCharEnum<MaterialStyleShaderSupportsDamageEnum>,
+}
+
 #[derive(Default, Debug, TagStructure)]
 #[data(size(0x88))]
 struct MaterialTag {
@@ -75,6 +146,10 @@ struct MaterialTag {
     material_parameters: FieldBlock<MaterialParameter>,
     #[data(offset(0x40))]
     postprocess_definition: FieldBlock<PostProcessDefinition>,
+    #[data(offset(0x6A))]
+    flags: FieldByteFlags<MaterialFlags>,
+    #[data(offset(0x74))]
+    style_info: FieldBlock<MaterialStyleInfo>,
 }
 
 fn main() -> infinite_rs::Result<()> {
@@ -94,12 +169,7 @@ fn main() -> infinite_rs::Result<()> {
             if tag.tag_group == "mat " {
                 let mut mat = MaterialTag::default();
                 tag.read_metadata(&mut mat)?;
-                for param in mat.postprocess_definition.elements.iter() {
-                    if param.textures.elements.is_empty() {
-                        continue;
-                    }
-                    println!("{:#?}", param.textures);
-                }
+                println!("{:#?}", mat.style_info);
             }
         }
     }