+++ infinite-rs/src/tag/zoneset/instance.rs
@@ -1,10 +1,9 @@
 //! Main zoneset structure.
 
 use byteorder::{ReadBytesExt, LE};
-use std::io::BufRead;
 
 use super::{instance_tag::TagZonesetInstanceHeader, tag::TagZonesetTag};
-use crate::common::extensions::{BufReaderExt, Readable};
+use crate::common::extensions::{BufReaderExt, Enumerable};
 use crate::Result;
 
 #[derive(Default, Debug)]
@@ -12,19 +11,16 @@ use crate::Result;
 pub(crate) struct TagZoneset {
     /// Header containing info on how many tags, footers, and parents to read.
     header: TagZonesetInstanceHeader,
-    /// List of tags with the size of `tag_count`.
+    /// List of tags with the size of [`tag_count`](`super::instance_tag::TagZonesetInstanceHeader::tag_count`).
     tags: Vec<TagZonesetTag>,
-    /// List of footer tags with the size of `footer_count`.
+    /// List of footer tags with the size of [`footer_count`](`super::instance_tag::TagZonesetInstanceHeader::footer_count`).
     footer_tags: Vec<TagZonesetTag>,
     /// List of 32-bit integers, unknown use.
     parents: Vec<i32>,
 }
 
-impl Readable for TagZoneset {
-    fn read<R>(&mut self, reader: &mut R) -> Result<()>
-    where
-        R: BufRead + BufReaderExt,
-    {
+impl Enumerable for TagZoneset {
+    fn read<R: BufReaderExt>(&mut self, reader: &mut R) -> Result<()> {
         self.header.read(reader)?;
         self.tags = reader.read_enumerable::<TagZonesetTag>(u64::from(self.header.tag_count))?;
         self.footer_tags =